requirements.txt
# Core Dependencies
discord.py==2.3.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
